backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asttokens==3.0.0
black==23.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
Django==4.2.7
django-cors-headers==4.3.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dj-database-url==2.1.0
python-decouple==3.8
whitenoise==6.6.0
executing==2.2.1
factory-boy==3.3.0
Faker==37.11.0
flake8==6.1.0
groq==0.4.1
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
ipython==8.17.2
isort==5.12.0
jedi==0.19.2
matplotlib-inline==0.2.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
pure_eval==0.2.3
pycodestyle==2.11.1
pydantic==2.12.3
pydantic_core==2.41.4
pyflakes==3.1.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
